backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
sqlalchemy
python-jose[cryptography]
passlib
email-validator
supabase
openai
pandas
openpyxl
python-multipart
requests
duckdb
xlrd
matplotlib
reportlab
python-docx
